backend/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.12.0
django-minio-storage==0.5.7
django-split-settings==1.3.2
django-stubs-ext==5.1.3
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.3.1
gunicorn==23.0.0
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
minio==7.2.15
msgpack==1.1.0
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.22.0
pyOpenSSL==25.0.0
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
service-identity==24.2.0
setuptools==76.1.0
six==1.17.0
sqlparse==0.5.3
transliterate==1.10.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uuid==1.30
uvicorn==0.34.0
zope.interface==7.2
# websockets==12.0
# wsproto==1.2.0